+++ src/race.js
@@ -14,22 +14,30 @@ const race = promises => new Promise((resolve, reject) => {
   const results = new Array(promises.length);
   let resolved = false;
   const unihandler = (idx, val, err) => {
+    // if already resolved, ignore further completions.
     if (resolved) {
       return;
     }
-    results[idx] = { type: val ? 'value' : 'error', payload: val || err };
+    // store the result of the promise
+    results[idx] = {
+      type: val ? 'value' : 'error',
+      payload: val || err,
+    };
 
+    // find the first successful result
     for (const r of results) {
       if (r === undefined) {
         return;
-      } else if (r.type === 'value') {
+      }
+      if (r.type === 'value') {
         resolve(r.payload);
         resolved = true;
         return;
       }
     }
 
-    reject(results.pop().payload);
+    // if all promises were rejected, reject with an array of errors.
+    reject(results.map(r => r.payload));
   };
 
   promises.forEach((p, idx) => {